# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem Edulite",
        "title": "Edulite API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5555",
    "basePath": "/",
    "paths": {
        "/siswa/{nisn}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil data siswa berdasarkan NISN",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Siswa"
                ],
                "summary": "Get siswa by NISN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NISN",
                        "name": "nisn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Siswa"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Siswa": {
            "type": "object",
            "properties": {
                "id_siswa": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "nisn": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}